Depricates the use of pxr_install_test_dir
Test directory installation is now handled by pxr_register_test.  The behavior of the `TESTENV` parameter to this function has been changed to ensure that an install rule is created for the required test directory and the correct directory is passed to the test runner without the test writer having to manually specify a destination.  Added a deprecation warning message to pxr_install_test_dir.  A very small percentage of tests require a specific destination path.  for those cases a `TESTENV_DEST` parameter is available to facilitate that requirement.

(Internal change: 2385894)